modify image upload button
