+++ app/assets/stylesheets/style.css
@@ -13,6 +13,8 @@ h3 {
 }
 p {font-size: 18px}
 
+.hidden {display: none;}
+
 /* header **************************************/
 header {
   height: 100px;
