+++ app/javascript/packs/application.js
@@ -9,6 +9,7 @@ require("@rails/activestorage").start()
 require("channels")
 require("../card")
 require("../preview")
+require("../preview_new")
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
