+++ app/javascript/preview_new.js
@@ -1,0 +1,45 @@
+if (document.URL.match( /services/ )) {
+  document.addEventListener('DOMContentLoaded', function(){
+    const ImageList = document.getElementById('image-list');
+
+    const createImageHTML = (blob) => {
+       // 画像を表示するためのdiv要素を生成
+      const img_up = document.getElementById('img-up');
+      const imageElement = document.createElement('div');
+      imageElement.setAttribute('class', "image-element")
+      let imageElementNum = document.querySelectorAll('.image-element').length
+
+      // 表示する画像を生成
+      const blobImage = document.createElement('img');
+      blobImage.setAttribute('src', blob);
+
+      // ファイル選択ボタンを生成
+      const inputHTML = document.createElement('input')
+      inputHTML.setAttribute('id', `service_image_${imageElementNum}`)
+      inputHTML.setAttribute('name', 'service[images][]')
+      inputHTML.setAttribute('type', 'file')
+      // inputHTML.setAttribute('class', "hidden")
+
+      // 生成したHTMLの要素をブラウザに表示させる
+      imageElement.appendChild(blobImage);
+      imageElement.appendChild(inputHTML);
+      ImageList.appendChild(imageElement);
+
+      img_up.appendChild(inputHTML);
+
+      inputHTML.addEventListener('change', (e) => {
+        file = e.target.files[0];
+        blob = window.URL.createObjectURL(file);
+        createImageHTML(blob)
+        inputHTML.setAttribute('class','hidden')
+      })
+    };
+
+    document.getElementById('service_image').addEventListener('change', function(e){
+      const file = e.target.files[0];
+      const blob = window.URL.createObjectURL(file);
+      createImageHTML(blob);
+      service_image.setAttribute('class','hidden')
+    });
+  });
+}
