+++ app/views/services/new.html.erb
@@ -23,12 +23,11 @@
         <%= f.collection_select(:category_id, Category.all, :id, :name, {} , {class:"category-select"}) %><br>
       </div>
 
-      <div class="field">
+      <div class="field" id="img-up">
         <p>画像</p>
         <%= f.file_field :images, name: 'service[images][]', id: 'service_image' %>
-        <div id="image-list"></div>
       </div>
-
+      <div id="image-list" class="d-flex flex-wrap"></div>
       <div class="field">
         <%= f.submit '作成' %>
       </div>
